Rename workspace_finder to workspace
