+++ ibazel/BUILD
@@ -49,7 +49,7 @@ go_library(
         "//ibazel/log:go_default_library",
         "//ibazel/output_runner:go_default_library",
         "//ibazel/profiler:go_default_library",
-        "//ibazel/workspace_finder:go_default_library",
+        "//ibazel/workspace:go_default_library",
         "//third_party/bazel/master/src/main/protobuf/blaze_query:go_default_library",
         "@com_github_fsnotify_fsnotify//:go_default_library",
     ],
@@ -68,7 +68,7 @@ go_test(
         "//bazel/testing:go_default_library",
         "//ibazel/command:go_default_library",
         "//ibazel/log:go_default_library",
-        "//ibazel/workspace_finder:go_default_library",
+        "//ibazel/workspace:go_default_library",
         "//third_party/bazel/master/src/main/protobuf/blaze_query:go_default_library",
         "@com_github_fsnotify_fsnotify//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
