+++ ibazel/output_runner/BUILD
@@ -21,7 +21,7 @@ go_library(
     visibility = ["//ibazel:__subpackages__"],
     deps = [
         "//ibazel/log:go_default_library",
-        "//ibazel/workspace_finder:go_default_library",
+        "//ibazel/workspace:go_default_library",
         "//third_party/bazel/master/src/main/protobuf/blaze_query:go_default_library",
     ],
 )
@@ -32,5 +32,5 @@ go_test(
     data = ["output_runner_test.json"],
     embed = [":go_default_library"],
     importpath = "github.com/bazelbuild/bazel-watcher/ibazel",
-    deps = ["//ibazel/workspace_finder:go_default_library"],
+    deps = ["//ibazel/workspace:go_default_library"],
 )
